Concluidas as funcionalidades de criar, remover, editar, buscar. buscar por categoria, buscar por subcategoria, buscar por promocao e buscar todas, referentes a Mercadoria. Concluído também as funcionalidades de upload e buscar imagens de mercadorias

+++ src/main/java/br/com/totemAutoatendimento/aplicacao/mercadoria/categoria/RemoverCategoria.java
@@ -1,19 +1,32 @@
 package br.com.totemAutoatendimento.aplicacao.mercadoria.categoria;
 
+import org.springframework.data.domain.Pageable;
+
+import br.com.totemAutoatendimento.dominio.exception.ViolacaoDeIntegridadeDeDadosException;
+import br.com.totemAutoatendimento.dominio.mercadoria.MercadoriaRepository;
 import br.com.totemAutoatendimento.dominio.mercadoria.categoria.Categoria;
 import br.com.totemAutoatendimento.dominio.mercadoria.categoria.CategoriaRepository;
 
 public class RemoverCategoria {
 
-private CategoriaRepository repository;
-	
-	public RemoverCategoria(CategoriaRepository repository) {
+	private CategoriaRepository repository;
+
+	private MercadoriaRepository mercadoriaRepository;
+
+	public RemoverCategoria(CategoriaRepository repository, MercadoriaRepository mercadoriaRepository) {
 		this.repository = repository;
+		this.mercadoriaRepository = mercadoriaRepository;
 	}
-	
+
 	public void executar(Long id) {
 		BuscarCategoria buscarCategoria = new BuscarCategoria(repository);
 		Categoria categoria = buscarCategoria.executar(id);
+		int quantidadeDeMercadoria = mercadoriaRepository.buscarPorCategoria(Pageable.unpaged(), categoria).getContent()
+				.size();
+		if (quantidadeDeMercadoria > 0) {
+			throw new ViolacaoDeIntegridadeDeDadosException(
+					"Impossível remover categoria pois existem " + quantidadeDeMercadoria + " mercadorias pertencentes a ela!");
+		}
 		repository.remover(categoria);
 	}
 }
